[REFACTOR] 컨텐츠 기능분리
