+++ src/components/detail/content/Contents.tsx
