+++ src/constants/global/querykeys.ts
@@ -1,6 +1,7 @@
 export const QUERY_KEYS = {
   KEYWORDLIST: "keywordList",
   POSTLIST: "postlist",
+  POST: "post",
   COMMENTS: "comments",
   ALERTSETTING: "alertSetting",
 };
